integration with MolGX provided as an extrernal package
