+++ extras_requirements.txt
@@ -1,3 +1,2 @@
 # extras requirements from custom pypi repositories
-AMD-Analytics==3.2.13
 cogmol_inference==0.6.0
