+++ requirements.txt
@@ -6,6 +6,7 @@ torchmetrics<0.7
 keybert==0.2.0
 minio==7.0.1
 modlamp>=4.0.0
+molgx>=0.22.0a0
 scikit-learn<0.24.0
 numpy>=1.16.5
 pytorch_lightning<=1.5.0
