+++ setup.cfg
@@ -65,7 +65,6 @@ gt4sd =
 [options.extras_require]
 extras =
     cogmol-inference
-    AMD-Analytics
 
 [flake8]
 max-line-length = 80
@@ -149,7 +148,7 @@ ignore_missing_imports = True
 [mypy-moses.*]
 ignore_missing_imports = True
 
-[mypy-AMD_Analytics.*]
+[mypy-molgx.*]
 ignore_missing_imports = True
 
 [mypy-paccmann_predictor.*]
