+++ src/gt4sd/extras/__init__.py
@@ -25,7 +25,6 @@
 # extras requirements
 EXTRAS_ENABLED: bool
 try:
-    import AMD_Analytics  # noqa: F401
     import cog  # noqa: F401
     import pag  # noqa: F401
